Add UserPolicy table
Changing this from InlinePolicy, which means we'll have seperate
tables for inline RolePolicy, and GroupPolicies. These all should
be the same but doing this to better align with the API here. This
also allows us to associate the correct identity with just the
GetUserPolicy response, the foriegn key just becomes UserName which
is what is provided in the API response for GetUserPoliicy.